backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.7.14
click==8.2.1
colorama==0.4.6
fastapi==0.116.1
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
multidict==6.6.3
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
yarl==1.20.1
# Additional dependencies for JWT and OAuth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
email-validator==2.1.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-api-python-client==2.178.0

# Document processing and vector embeddings
langchain==0.3.27
langchain-community==0.3.27
langchain-text-splitters==0.3.9
langchain-core==0.3.72
numpy==1.26.4
pypdf==4.2.0
docx2txt==0.8
# Supabase for database, storage, and vector embeddings
supabase==2.18.1

# Google AI for LangChain
langchain-google-genai
google-generativeai
openai==1.58.1

# Testing
pytest
pytest-asyncio
httpx
asyncio
greenlet